requirements.txt
streamlit>=1.22.0
sentence-transformers>=2.2.2
numpy>=1.22.0
scikit-learn>=1.1.0
torch>=1.12.0
requests>=2.28.1
python-dotenv>=0.19.0
nltk>=3.8.0
faiss-cpu>=1.7.3
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

